+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Features
 --------
 * Implement serializers for the Graph String API (PYTHON-778)
 * Provide deserializers for GraphSON types (PYTHON-782)
+* Add Graph DurationType support (PYTHON-607)
 
 Merged from OSS master (pre-3.11):